+++ external/o2/src/o2.h
@@ -103,7 +103,7 @@ public:
     QString refreshToken();
 
     /// Get token expiration time (seconds from Epoch).
-    int expires();
+    long long expires();
 
 public Q_SLOTS:
     /// Authenticate.
@@ -159,7 +159,7 @@ protected:
     void setRefreshToken(const QString &v);
 
     /// Set token expiration time.
-    void setExpires(int v);
+    void setExpires(long long v);
 
     virtual QNetworkAccessManager *getManager();
