+++ src/analysis/processing/qgsalgorithmpackage.cpp
@@ -130,7 +130,7 @@ QVariantMap QgsPackageAlgorithm::processAlgorithm( const QVariantMap &parameters
 
   bool errored = false;
 
-  QgsProcessingMultiStepFeedback multiStepFeedback( mLayers.size(), feedback );
+  QgsProcessingMultiStepFeedback multiStepFeedback( static_cast<int>( mLayers.size() ), feedback );
 
   QStringList outputLayers;
   int i = 0;
