+++ src/app/3d/qgspointcloud3dsymbolwidget.h
@@ -45,7 +45,7 @@ class QgsPointCloud3DSymbolWidget : public QWidget, private Ui::QgsPointCloud3DS
     void setPointBudget( double budget );
     double pointBudget() const;
 
-    void setPointCloudSize( int size );
+    void setPointCloudSize( long long size );
 
     void connectChildPanels( QgsPanelWidget *parent );
